Use a fake @wechange mail for rocketchat for unverified users
Rocketchat is broken for their "unverified" setting and will still sometimes send out emails to those users. In addition, on an OAuth login, rocket will set a user to verified always, which is terrible for us. Saving back the verified=False state after an OAuth login will cause the user account to see a "re-enter your password" which we can't allow.